Add files via upload
fun fact about this is when we use b= 8342093840949 & c= 293023902 we get x as '0.0' and y as '-b' because when calculating disc b*b is too large as compared to 4*c making disc value as b therefore giving us x ( i.e. -b + b now ) as 0 and y as -b.